Move `checkEnabledExchanges` to SettingsActions
