+++ src/components/Main.tsx
@@ -6,10 +6,9 @@ import * as React from 'react'
 import { Platform } from 'react-native'
 import { AirshipToast } from 'react-native-airship'
 
-import { checkEnabledExchanges } from '../actions/CryptoExchangeActions'
 import { getDeviceSettings } from '../actions/DeviceSettingsActions'
 import { logoutRequest } from '../actions/LoginActions'
-import { showReEnableOtpModal } from '../actions/SettingsActions'
+import { checkEnabledExchanges, showReEnableOtpModal } from '../actions/SettingsActions'
 import { CryptoExchangeScene as CryptoExchangeSceneComponent } from '../components/scenes/CryptoExchangeScene'
 import { HomeSceneUi4 as HomeSceneUi4Component } from '../components/ui4/scenes/HomeSceneUi4'
 import { ENV } from '../env'
